refactor(tests/e2e): group alerts specs under tests/alerts/
